moved `getBalanceUntilLockTime` to `OutputQueries`
