+++ app/src/main/scala/org/alephium/explorer/persistence/queries/TransactionQueries.scala
@@ -351,31 +351,6 @@ object TransactionQueries extends StrictLogging {
         (total.getOrElse(U256.Zero), locked.getOrElse(U256.Zero))
     }
 
-  def getBalanceActionOption(address: Address)(
-      implicit ec: ExecutionContext): DBActionR[(Option[U256], Option[U256])] =
-    getBalanceUntilLockTime(
-      address  = address,
-      lockTime = TimeStamp.now()
-    )
-
-  def getBalanceUntilLockTime(address: Address, lockTime: TimeStamp)(
-      implicit ec: ExecutionContext): DBActionR[(Option[U256], Option[U256])] =
-    sql"""
-      SELECT sum(outputs.amount),
-             sum(CASE
-                     WHEN outputs.lock_time is NULL or outputs.lock_time < ${lockTime.millis} THEN 0
-                     ELSE outputs.amount
-                 END)
-      FROM outputs
-               LEFT JOIN inputs
-                         ON outputs.key = inputs.output_ref_key
-                             AND inputs.main_chain = true
-      WHERE outputs.spent_finalized IS NULL
-        AND outputs.address = $address
-        AND outputs.main_chain = true
-        AND inputs.block_hash IS NULL;
-    """.as[(Option[U256], Option[U256])].exactlyOne
-
   // switch logger.trace when we can disable debugging mode
   protected def debugShow(query: slickProfile.ProfileAction[_, _, _]) = {
     print(s"${query.statements.mkString}\n")
